+++ packages/md-enhance/__tests__/unit/sup.spec.ts
@@ -33,7 +33,7 @@ describe("superscript", () => {
     );
   });
 
-  it("Should handle mutiple '\\'", () => {
+  it("Should handle multiple '\\'", () => {
     expect(markdownIt.render(`^foo\\\\\\\\\\\\\\ bar^`)).toEqual(
       "<p><sup>foo\\\\\\ bar</sup></p>\n"
     );