+++ CHANGELOG.md
@@ -1290,7 +1290,7 @@ Since `vuepress` drop support for the `babel` way, `search` and `docsearch` opti
 
 ### Bug Fixes
 
-- **components:** allow mutiple $page ([1f9b8f7](https://github.com/vuepress-theme-hope/vuepress-theme-hope/commit/1f9b8f731d0d47c407774cec0d586df3bcb2de20))
+- **components:** allow multiple $page ([1f9b8f7](https://github.com/vuepress-theme-hope/vuepress-theme-hope/commit/1f9b8f731d0d47c407774cec0d586df3bcb2de20))
 - **components:** improve a11y ([cc506bc](https://github.com/vuepress-theme-hope/vuepress-theme-hope/commit/cc506bc9651843dddbdd49fa4db39051350a52c9))
 - style fixes ([c9430f5](https://github.com/vuepress-theme-hope/vuepress-theme-hope/commit/c9430f59f8a384e8fd90bc68a23307950fdea56d))
 - **theme:** fix navbar style ([92503bd](https://github.com/vuepress-theme-hope/vuepress-theme-hope/commit/92503bd7753ace84054ae700b22ee9701ca59ba6))